pyproject.toml
[project]
name = "aneval"
version = "0.1.0"
description = ""
authors = [
    {name = "DimitarAtanassov",email = "dimitar789558@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "pydantic (>=2.12.3,<3.0.0)",
    "streamlit (>=1.35.0,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "yfinance (>=0.2.66,<0.3.0)",
    "openai (>=2.8.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "aneval", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
